pyproject.toml
[tool.poetry]
name = "tgfloofbot"
version = "0.0.0"
description = ""
authors = []
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
pydantic = "^1.7.3"
SQLAlchemy = "^1.3.23"
alembic = "^1.5.5"
pytest = {version = "^6.2.2", optional = true}
mkdocs = {version = "^1.1.2", optional = true}
mkdocs-material = {version = "^7.0.0", optional = true}
mypy = {version = "^0.812", optional = true}
black = {version = "^20.8b1", optional = true}
pytest-mypy = {version = "^0.8.0", optional = true}
pytest-black = {version = "^0.3.12", optional = true}
flake8 = {version = "^3.8.4", optional = true}
python-telegram-bot = "13.3"

[tool.poetry.extras]
lint = ["mypy", "black", "flake8"]
test = ["pytest", "pytest-mypy", "pytest-black", "mkdocs", "mkdocs-material"]
docs = ["mkdocs", "mkdocs-material"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
